code/svb-docker/python/pyproject.toml
[tool.poetry]
name = "svb"
version = "0.1.0"
description = ""
authors = ["John McNelly <jkailimcnelly@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
python-escpos = "^2.2.0"
qrcode = "^7.4.2"
pillow = "^10.1.0"
pyyaml = "^6.0.1"
pytest-django = "^4"
pytest-env = "^0"
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
